fix create and read orders requests

+++ src/main/java/com/example/service/ShopService.java
@@ -1,6 +1,8 @@
 package com.example.service;
 
 
+import com.example.exception.ShopException;
+import com.example.exception.ShopExceptionTitle;
 import com.example.model.Customer;
 import com.example.model.Product;
 import com.example.model.ShopOrder;
@@ -12,6 +14,7 @@ import com.example.tools.StringJsonConverter;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -19,12 +22,11 @@ import java.util.UUID;
 @Service
 @AllArgsConstructor
 public class ShopService {
+    private final StringJsonConverter stringJsonConverter = new StringJsonConverter();
     private CustomerRepository customerRepository;
     private ProductRepository productRepository;
     private ShopOrderRepository shopOrderRepository;
-    private final StringJsonConverter stringJsonConverter = new StringJsonConverter();
 
-    //TODO handle exceptions for crud operation product
     public String createProduct(String jsonProduct) {
         Product product = stringJsonConverter.toObject(jsonProduct, Product.class);
         Product savedProduct = productRepository.save(product);
@@ -36,8 +38,8 @@ public class ShopService {
         if (optionalProduct.isPresent()) {
             return stringJsonConverter.toString(optionalProduct.get());
         }
-        // TODO add throw Exception
-        throw new RuntimeException("product with id not exist");
+
+        throw new ShopException(ShopExceptionTitle.PRODUCT_WITH_SUCH_ID_DOESNT_EXIST);
     }
 
     public String readAllProducts() {
@@ -47,19 +49,23 @@ public class ShopService {
 
 
     public String updateProduct(String jsonProduct) {
-        Product product = stringJsonConverter.toObject(jsonProduct, Product.class); // TODO what if product has got with null id?
+        Product product = stringJsonConverter.toObject(jsonProduct, Product.class);
         Optional<Product> optionalProduct = productRepository.findById(product.getProductId());
         if (optionalProduct.isPresent()) {
             Product savedProduct = productRepository.save(product);
             return stringJsonConverter.toString(savedProduct);
         }
-
-        // TODO add throw Exception
-        throw new RuntimeException("product with id not exist");
+        throw new ShopException(ShopExceptionTitle.PRODUCT_WITH_SUCH_ID_DOESNT_EXIST);
     }
 
     public void deleteProduct(UUID id) {
-        productRepository.deleteById(id);
+
+        if (productRepository.existsById(id)) {
+            productRepository.deleteById(id);
+        } else {
+            throw new ShopException(ShopExceptionTitle.PRODUCT_WITH_SUCH_ID_DOESNT_EXIST);
+        }
+
     }
 
     public String createCustomer(String jsonCustomer) {
@@ -69,32 +75,50 @@ public class ShopService {
     }
 
 
-    //TODO add checking count of products
     public String createOrder(String jsonOrderDto) {
         ShopOrderDTO shopOrderDTO = stringJsonConverter.toObject(jsonOrderDto, ShopOrderDTO.class);
         // get consumer
         Optional<Customer> optionalCustomer = customerRepository.findById(shopOrderDTO.getCustomerId());
         if (optionalCustomer.isEmpty()) {
-            // TODO add throw custom Exception
-            throw new RuntimeException("Customer with id not exist");
+            throw new ShopException(ShopExceptionTitle.CUSTOMER_WITH_SUCH_ID_DOESNT_EXIST);
         }
+
         // get products
         List<Product> products = productRepository.findAllById(shopOrderDTO.getProductsId());
         if (products.isEmpty()) {
-            // TODO add throw custom Exception
-            throw new RuntimeException("Cannot create order without products");
+            throw new ShopException(ShopExceptionTitle.EMPTY_PRODUCTS_LIST);
         }
-        //save order
+
         ShopOrder order = new ShopOrder(shopOrderDTO);
+
+        List<Product> acceptableProducts = products.stream()
+                .filter(product -> {
+                    return product.getQuantityInStock() > product.getOrders().size();
+                }).toList();
+
+        acceptableProducts.forEach(product -> {
+            order.getProducts().add(product);
+            product.getOrders().add(order);
+        });
+        productRepository.saveAll(acceptableProducts);
+
+        //save order
         order.setCustomer(optionalCustomer.get());
-        order.setProducts(products);
+        order.setProducts(acceptableProducts);
+        BigDecimal totalPrice = acceptableProducts.stream().map(Product::getPrice).reduce(BigDecimal.ZERO, BigDecimal::add);
+        order.setTotalPrice(totalPrice);
         ShopOrder savedOrder = shopOrderRepository.save(order);
         return stringJsonConverter.toString(savedOrder);
     }
 
 
     public String readOrder(UUID id) {
-        return "need implement";
+        Optional<ShopOrder> optionalShopOrder = shopOrderRepository.findById(id);
+
+        if (optionalShopOrder.isPresent()) {
+            return stringJsonConverter.toString(optionalShopOrder.get());
+        }
+        throw new ShopException(ShopExceptionTitle.ORDER_WITH_SUCH_ID_DOESNT_EXIST);
     }
 }
